requirements.txt
flask==2.0.1
firebase-admin==6.2.0
python-dotenv==0.19.0
gunicorn==20.1.0